+++ src/main/java/com/notenoughmail/kubejs_tfc/block/internal/AbstractCropBlockBuilder.java
@@ -5,7 +5,6 @@ import com.notenoughmail.kubejs_tfc.block.sub.DeadCropBlockBuilder;
 import com.notenoughmail.kubejs_tfc.util.DataUtils;
 import com.notenoughmail.kubejs_tfc.util.RegistryUtils;
 import dev.latvian.mods.kubejs.block.BlockBuilder;
-import dev.latvian.mods.kubejs.block.BlockItemBuilder;
 import dev.latvian.mods.kubejs.block.SeedItemBuilder;
 import dev.latvian.mods.kubejs.client.ModelGenerator;
 import dev.latvian.mods.kubejs.client.VariantBlockStateGenerator;
@@ -62,15 +61,11 @@ public abstract class AbstractCropBlockBuilder extends ExtendedPropertiesBlockBu
         renderType("cutout");
         productItem = null;
         RegistryUtils.hackBlockEntity(TFCBlockEntities.CROP, this);
+        noItem();
     }
 
     protected boolean hasProduct() {
         return true;
-    }
-
-    @Override
-    public BlockBuilder item(@Nullable Consumer<BlockItemBuilder> i) {
-        return this;
     }
 
     @Info(value = "Determines how many growth stages the crop will have")