+++ src/main/java/com/notenoughmail/kubejs_tfc/event/RegisterItemStackModifierEventJS.java
@@ -17,7 +17,7 @@ import java.util.function.Supplier;
 public class RegisterItemStackModifierEventJS extends EventJS {
 
     @Deprecated(forRemoval = true, since = "1.2.0")
-    @Info(value = "Creates a new item stack modifier with the given id and function", params = {
+    @Info(value = "Deprecated, please use `simple`, `withInput`, or `withInventory` to register a modifier instead. Creates a new item stack modifier with the given id and function", params = {
             @Param(name = "id", value = "The registry name of the modifier"),
             @Param(name = "applicator", value = "The function which will be applied when the modifier is applied to the stack"),
             @Param(name = "dependsOnInput", value = "Sets if the modifier is dependent on the input item, should be true if the input stack or the crafting container/inputs of the applicator is used")
@@ -51,7 +51,7 @@ public class RegisterItemStackModifierEventJS extends EventJS {
     }
 
     @Deprecated(forRemoval = true, since = "1.2.0")
-    @Info(value = "Returns a supplier for the current crafting container. Only available during advanced shaped and shapeless crafting recipes. Deprecated, use `getCraftingInput()` instead")
+    @Info(value = "Deprecated, use `withInventory` to register a modifier with access to the inventory instead. Returns a supplier for the current crafting container. Only available during advanced shaped and shapeless crafting recipes")
     @Generics(CraftingContainer.class)
     public Supplier<@Nullable CraftingContainer> getCraftingContainer() {
         return RecipeHelpers::getCraftingContainer;